get-live-campus-busyness/requirements.txt
functions-framework==3.*
firebase-admin==6.*
requests==2.*
google-cloud-logging==3.*